% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_p.R
\name{compare_p}
\alias{compare_p}
\title{Compare traditional, max, and second generation P-values across studies}
\usage{
compare_p(data, delta_a, delta_b, h0, plot = TRUE)
}
\arguments{
\item{data}{Data frame with two variables \code{xbar} and \code{se}.}

\item{delta_a}{Numeric. Lower bound of the indifference zone}

\item{delta_b}{Numeric. Upper bound of the indifference zone}

\item{h0}{Numeric. Value of the null hypothesis}

\item{plot}{Logical. Whether to produce a plot of the study values with the
indifference zone (default = \code{TRUE}).}
}
\value{
data frame with the following columns:
\itemize{
\item xbar: Mean corresponding to the input \code{xbar} value
\item se: Standard error corresponding to the input \code{se} value
\item lb: Lower bound of the effect (95% CI)
\item ub: Upper bound of the effect (95% CI)
\item p_old: Traditional P-value
\item p_max: Maximum P-value over the interval null
\item p_new: Second Generation P-value
}

If \code{plot = TRUE} (default), will also output a plot of the studies with the
indifference zone.
}
\description{
Compare traditional, max, and second generation P-values across studies
}
\examples{
## Figure 2 from the second generation P-value paper
data <- data.frame(
  xbar = c(146, 145.5, 145, 146, 144, 143.5, 142, 141),
  se = c(0.5, 0.25, 1.25, 2.25, 1, 0.5, 1, 0.5)
  )
compare_p(data, delta_a = 144, delta_b = 148, h0 = 146)
}
