% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leukdata.R
\docType{data}
\name{leukdata}
\alias{leukdata}
\title{Leukemia Data}
\format{A matrix with 513216 elements}
\usage{
leukdata
}
\description{
Data from the ALL-AML Leukemia Microarray Study used to generate
Figures 3 and 4 and Table 2 in the Second-generation p-values:
Improved rigor, reproducibility, & transparency in statistical analyses
manuscript. A microarray analysis was conducted on 7128 candidate genes and
two-group t-tests were performed on the cleaned and standardized log10
expression levels of each gene.
}
\references{
Matejka, J., & Fitzmaurice, G. (2017).
Golub TR, Slonim DK, Tamayo P, Huard C, Gaasenbeek M, Mesirov JP, et. al. (1999).
Molecular classification of cancer: class discovery and class prediction by gene
expression monitoring. Science, 286(5439): 531–537. pmid:10521349
}
\keyword{datasets}
