% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_delta.R
\name{p_delta}
\alias{p_delta}
\title{Second Generation P-value}
\usage{
p_delta(lb, ub, delta_a, delta_b)
}
\arguments{
\item{lb}{Numeric. Lower bound of the observed effect}

\item{ub}{Numeric. Upper bound of the observed effect}

\item{delta_a}{Numeric. Lower bound of the indifference zone}

\item{delta_b}{Numeric. Upper bound of the indifference zone}
}
\value{
Numeric. Second generation p-value for the observed effect.
}
\description{
Second Generation P-value
}
\examples{
## Second Generation P-value for a confidence interval of (140, 145) and an
## indifference zone of (144, 148)
second_gen_pvalue(140, 145, 144, 148)
}
