% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_curves.R
\name{power_curves}
\alias{power_curves}
\title{Power Curves varying delta}
\usage{
power_curves(n, delta, truth = "null")
}
\arguments{
\item{n}{Numeric. The sample size}

\item{delta}{Numeric vector. The width of the indifference zone in SD units. Should be between
-3 and 3, as this is the range of the \eqn{\theta} space (and likely much smaller).}

\item{truth}{Character. True data generating mechanism - must be one of the following:
\itemize{
\item null
\item alternative
\item inconclusive
}}
}
\value{
A plot of the power curves
}
\description{
The data for this function are generated where the standardized \eqn{\theta} is from -3 to 3, so
\eqn{\delta} values, values for the width of the indifference zone, should be certainly be smaller.
}
\examples{
## Figure S3 from the second generation P-value paper. The relationship between
## the probability that p_delta = 0 varying delta.

power_curves(n = 10, delta = c(0, 1/30, 1/2, 1), truth = "null")

## Figure S5 from the second generation P-value paper. The relationship
## between probability of data supported compatibility 517 with the null hypothesis,
## and various deltas

## n = 40 (Figure S5 left)
power_curves(n = 40, delta = c(0, 1/30, 1/2, 1), truth = "alternative")
## n = 200 (Figure S5 right)
power_curves(n = 200, delta = c(0, 1/30, 1/2, 1), truth = "alternative")

## Figure S6 from the second generation P-value paper. The relationship between
## the probability of an inconclusive result and various deltas.

## n = 20 (Figure S6 left)
power_curves(n = 20, delta = c(0, 1/30, 1/2, 1), truth = "inconclusive")
## n = 200 (Figure S6 right)
power_curves(n = 200, delta = c(0, 1/30, 1/2, 1), truth = "inconclusive")
}
