% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_curves.R
\name{power_inconclusive}
\alias{power_inconclusive}
\title{Probability that \eqn{0 < p_\delta < 1}{0 < p\delta < 1}}
\usage{
power_inconclusive(delta, n, theta)
}
\arguments{
\item{delta}{Numeric vector. The width of the indifference zone in SD units.}

\item{n}{Numeric. The sample size.}

\item{theta}{The "true" effect size.}
}
\value{
Data frame with the following columns:
\itemize{
\item delta: The width of the simulated indifference zone
\item n: The sample size
\item theta: the "true" effect size
\item p: The probability \eqn{0 < p_\delta < 1}{0 < p\delta < 1}
}
}
\description{
Probability that \eqn{0 < p_\delta < 1}{0 < p\delta < 1}
}
\examples{
power_inconclusive(
   delta = c(0, 1/30, 1/2, 1),
   n = 33,
   theta = seq(-3, 3, 0.01)
   )

}
