% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_curves.R
\name{power_curves}
\alias{power_curves}
\title{Power Curves varying delta}
\usage{
power_curves(delta, n, theta = seq(-3, 3, 0.001), prob = "null")
}
\arguments{
\item{delta}{Numeric vector. The width of the indifference zone in SD units.}

\item{n}{Numeric. The sample size.}

\item{theta}{Numeric vector. The "true" effect size (i.e., the default is \code{seq(-3, 3, 0.001}).}

\item{prob}{Character. The probability of interest (i.e., if you are interested in the probability that
\eqn{p_\delta = 0}{p\delta = 0} use \code{null}). Must be one of the following:
\itemize{
\item null
\item alternative
\item inconclusive
}}
}
\value{
A plot of the power curves
}
\description{
Power Curves varying delta
}
\examples{
## Figure S3 from the second generation P-value paper. The relationship between
## the probability that p_delta = 0 varying delta.

power_curves(delta = c(0, 1/30, 1/2, 1), n = 10, prob = "null")

## Figure S5 from the second generation P-value paper. The relationship
## between probability of data supported compatibility with the null hypothesis,
## and various deltas

## n = 40 (Figure S5 left)
power_curves(delta = c(0, 1/30, 1/2, 1), n = 40, prob = "alternative")
## n = 200 (Figure S5 right)
power_curves(delta = c(0, 1/30, 1/2, 1), n = 200, prob = "alternative")

## Figure S6 from the second generation P-value paper. The relationship between
## the probability of an inconclusive result and various deltas.

## n = 20 (Figure S6 left)
power_curves(delta = c(0, 1/30, 1/2, 1), n = 20, prob = "inconclusive")
## n = 200 (Figure S6 right)
power_curves(delta = c(0, 1/30, 1/2, 1), n = 200, prob = "inconclusive")
}
